requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0
xgboost>=2.0.0
lightgbm>=4.0.0
optuna>=3.4.0
shap>=0.44.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
joblib>=1.3.0
mlflow>=2.9.0
pytest>=7.4.0
